Moved UserRole model
