+++ app/db/base.py
@@ -3,4 +3,5 @@
 from app.business_partner.models import *  # noqa
 from app.users.models import *
 from app.auth.models import *
+from app.roles.models import *
 from app.db.base_class import Base
